+++ open_alchemy/schemas/validation/__init__.py
@@ -1,5 +1,4 @@
 """Schema validation pre-processor."""
 
 from . import model
-from . import relationship
-from . import simple
+from . import property_
