+++ open_alchemy/schemas/validation/property_/__init__.py
@@ -1,0 +1,4 @@
+"""Validation for properties."""
+
+from . import relationship
+from . import simple
